[crypto/kats] Connect KATs
Connect the KATs to the cryptographic calls in the cryptolib.
Create a functest to run all kats independently.
